Add jmri.util.TypeConversionUtil
jmri.util.TypeConversionUtil converts between different types.

This commit also changes AnalogAction and AnalogExpression to use double instead of float.

+++ java/src/jmri/jmrit/logixng/AnalogAction.java
@@ -13,8 +13,8 @@ public interface AnalogAction extends NamedBean, Base {
      * Set an analog value.
      * 
      * @param value the value. The male socket that holds this action ensures
-     * that this value is not Float.NaN or an infinite value.
+     * that this value is not Double.NaN or an infinite value.
      */
-    public void setValue(float value);
+    public void setValue(double value);
     
 }
