+++ java/src/jmri/jmrit/logixng/MaleAnalogActionSocket.java
@@ -9,11 +9,11 @@ public interface MaleAnalogActionSocket
     /**
      * {@inheritDoc}
      * <P>
-     * This method must ensure that the value is not a Float.NaN, negative
+     * This method must ensure that the value is not a Double.NaN, negative
      * infinity or positive infinity. If that is the case, it must throw an
      * IllegalArgumentException before checking if an error has occured.
      */
     @Override
-    public void setValue(float value);
+    public void setValue(double value);
 
 }
